Add some no-op mocks
This commit adds all of the basic pipeline steps which can more or
less be mocked as no-ops. Note that stash/unstash appear here because
we are operating under the assumption that different stages of a build
will not execute on different build nodes, as can happen on a live
Jenkins bulid.